Added apps grouping in testing history (#47)
