+++ src/store/slices/testingHistory.slice.ts
@@ -6,7 +6,7 @@ import { isAnyTaskFailure } from "pages/certification/Certification.helper";
 
 import type { Run } from 'components/CreateCertificate/CreateCertificate';
 
-interface Certificate {
+export interface Certificate {
   transactionId: string;
 }
 
@@ -31,7 +31,7 @@ export const fetchHistory = createAsyncThunk("fetchHistory", async (payload: {},
   }
 });
 
-export const getRowStatus = createAsyncThunk("getRowStatus", async (payload: { index: number, runId: string }, thunkApi) => {
+export const getRowStatus = createAsyncThunk("getRowStatus", async (payload: { runId: string }, thunkApi) => {
   try {
     const response = await fetch<any>(thunkApi, { method: 'GET', url: `/run/${payload.runId}` });
     const state = response.data.hasOwnProperty("state") ? response.data.state : "";
@@ -45,7 +45,7 @@ export const getRowStatus = createAsyncThunk("getRowStatus", async (payload: { i
       const isComplete = isUnitTestSuccess && !isAnyTaskFailure(resultJson);
       status = isComplete ? 'succeeded' : 'failed';
     }
-    return { index: payload.index, status };
+    return { runId: payload.runId, status };
   } catch (e: any) {
     return thunkApi.rejectWithValue(e.response.data);
   }
@@ -92,7 +92,7 @@ const testingHistorySlice = createSlice({
       })
       .addCase(getRowStatus.fulfilled, (state, actions) => {
         state.history = state.history.map(
-          (row, index) => index !== actions.payload.index ? row : ({
+          row => row.runId !== actions.payload.runId ? row : ({
             ...row, runStatus: actions.payload.status as any
           })
         );
